Update scipy.cluster.vq.kmeans call
Cast array to numpy float type.

+++ photomosaic.py
@@ -87,7 +87,7 @@ def dominant_color(img, clusters=5, size=50):
     img.thumbnail((size, size))
     imgarr = scipy.misc.fromimage(img)
     imgarr = imgarr.reshape(scipy.product(imgarr.shape[:2]), imgarr.shape[2])
-    colors, dist = vq.kmeans(imgarr, clusters)
+    colors, dist = vq.kmeans(imgarr.astype(np.float), clusters)
     vecs, dist = vq.vq(imgarr, colors)
     counts, bins = scipy.histogram(vecs, len(colors))
     dominant_color = colors[counts.argmax()]
